backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
chromadb==0.5.23
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
huggingface-hub==0.26.5
lxml==5.3.0
markdownify==0.13.1
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==12.1.1
